Add a callback parameter to Hoptoad.notify
Having a callback is useful to shut down a node app after an
uncaughtException, while making sure Hoptoad has received the error.
Shutting down (and restarting) is usually a good idea since exceptions
cause the call stack to drop and most libraries will not take kindly
to sudden interruption. That is, their state will become inconsistent
and future calls to them might produce unpredictable results.

This patch does not include a test since JSpec seems to put plenty
of emphasis on cute syntax sugar, but does very little in the way of
advanced stubbing, faking and mocking.

I did consider bringing in gently as a dependency, but since I was
not sure if that would increase the chances of this getting merged
I just went ahead and implemented this the old school way.

Not happy with it, but I can't justify spending more time on it right
now.